requirements.txt
Django==5.1.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.4.0
gunicorn==22.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
redis==5.0.3
drf-spectacular[sidecar]==0.28.0
